:white_check_mark: Separates each ui test inside blocks
